feat: repeating event with columns
There are common overlaps with footnotes.
This looks weird when small fragments of a repeating event leak below.

Now a repating event can define appearsIn columns that are respected.
Note that we can use also columns not defined for that day (but that are used for other days).
This has the implication that a repeating event could be duplicated if the columns are not adjacent, or even left out completely if the columns do not exist for that day.

As a drawback we must give appersIn for these events with possibly many columns (but still only once since its repeating :)

when appearsIn is left empty, we still print them full width